oauth: Add HMR
